requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Type hints
pydantic>=2.0.0

# LLM and PDF processing
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# PDF processing
pypdf>=5.0.0

# Vector database
chromadb>=0.5.0
langchain-chroma>=0.1.0

# Text splitting
langchain-text-splitters>=0.3.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# UI
gradio>=4.0.0

